+++ evernote-sdk-ios/3rdParty/NSString+URLEncoding/NSString+URLEncoding.m
@@ -70,11 +70,11 @@
                              @"%23", @"%21", @"%27",
                              @"%28", @"%29", @"%2A", @"+", nil];
 	
-	int len = [escapeChars count];
+	NSUInteger len = [escapeChars count];
 	
 	NSMutableString *temp = [self mutableCopy];
 	
-	int i;
+	NSUInteger i;
 	for(i = 0; i < len; i++)
 	{
